Tweaked the orientation change hack
The problem with the page rotation in older versions of Silica was
that the page width was changing with a delay, causing visible and
unpleasant layout changes.

When the support for cutout was introduced, this hack started to
break the landscape layout (with cutout enabled, the width of the
page in landscape is smaller than the screen height) and at the same
time, the old unpleasant rotation effects seems to have gone away.

Therefore the absence of the cutout property is used as the criteria
for applying the hack.